Fix new clippy lints for Rust 1.66
- https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
- https://rust-lang.github.io/rust-clippy/master/index.html#manual_clamp
- https://rust-lang.github.io/rust-clippy/master/#suboptimal_flops

+++ src/asm/shared/transform/inverse.rs
@@ -73,7 +73,7 @@ pub fn call_inverse_hbd_func<T: Pixel>(
   unsafe {
     func(
       output.data_ptr_mut() as *mut _,
-      T::to_asm_stride(output.plane_cfg.stride) as isize,
+      T::to_asm_stride(output.plane_cfg.stride),
       copied.data.as_mut_ptr() as *mut _,
       eob as i32 - 1,
     );
